requirements.txt
absl-py==0.8.1
astor==0.8.0
bert-serving-client==1.9.8
gast==0.3.2
grpcio==1.25.0
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.17.3
protobuf==3.10.0
PyYAML==5.1.2
pyzmq==18.1.0
scipy==1.3.1
six==1.13.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.16.0
